Merge branch 'feature/add-s3-support' into 'main'
* fix incorrect bucket name addressing in DeleteModal

* Apply 3 suggestion(s) to 3 file(s)


* Apply 1 suggestion(s) to 1 file(s)


* Cherry-pick merge branch 'bugfix/project-level-upload' into feature/add-s3-support

* fix upload and download progress and cancellation

* throttle progress updates

* Initial s3 bucket copy implementation

* Use bucket name validation in Copy modal. Don't add spaces to autosuggested name.

* Comment out updating tags when buckets and objects are updated

* Comment out tag input on bucket copy

* Comment out tagging and search/filtering for buckets and objects

* Comment out tag input on bucket creation

* Match comment line length to code line length in eslint rules

* Get size and object count for own buckets

* Fix bucket deletion

* Apply 2 suggestion(s) to 1 file(s)


* Temporarily bypass CI test stage when merging to s3 feature branch

* Update bucket validation for s3. Check is last bucket name char is allowed.

* Update .env.test

* fix sharing permissions

* skip acl sync

* add s3 endpoint envar to default environment configuration

* revert to debian oldstable in submodule dev branch

* remove swift dependencies from bucket operations and sharing

* fix multipart upload issues

* fix incorrect import

* add s3 download worker socket

* fix incorrect file extension breaking decryption

* fix multipart upload issues

* use s3 for creating and listing folders

* host ui behind dockerized dev stack

* use default ports for sd-connect and rgw proxies

* add dev environment shorthands to makefile

* fix firefox incorrect policy causing cert issues

* add programmatically generated dev ca and browser containers with trusted CAs

* add initial makefile for development resources

* add initial s3 upload and download

* add container publish client

* remove legacy aws-sdk and minio

* allow using the S3 api in the frontend

* proxy keystone-swift through nginx to enable TLS

* support retrieving ec2 credentials from the backend

Closes #1253
See merge request https://gitlab.ci.csc.fi/sds-dev/sd-connect/swift-browser-ui/-/merge_requests/445

Approved-by: Timo Nurminen <tnurmine@csc.fi>
Merged by Sampsa Penna <sapenna@csc.fi>

+++ swift_browser_ui_frontend/build_wasm.sh
@@ -2,22 +2,44 @@
 
 set -e
 
-[ ! -x "$(command -v docker)" ] && echo "Docker is used to build the javascript WebAssembly dependencies, but it's not installed." && exit;
-if ! docker version > /dev/null 2>&1; then echo "Docker is installed, but it seems like there's an error."; exit 1; fi
-
 SCRIPT="$(realpath $0)"
 SCRIPT_ROOT=$(dirname "$SCRIPT")
-WASM_ROOT="${SCRIPT_ROOT}"/wasm/
+WASM_ROOT="${SCRIPT_ROOT}"/wasm
 
-docker run --rm -it --mount type=bind,source="${WASM_ROOT}",target=/src/ ghcr.io/cscfi/docker-emscripten-crypt4gh clean
+buildWasm() {
+    mkdir "${WASM_ROOT}/build"
+    (
+        cd $SCRIPT_ROOT
+        npx webpack --config "${WASM_ROOT}/wasm-webpack.config.js"
+        cd $WASM_ROOT
+    )
 
-mkdir "${WASM_ROOT}/build"
-(
+    if [ "$1" = "docker" ]; then
+        docker run --rm -it --mount type=bind,source="${WASM_ROOT}",target=/src/ ghcr.io/cscfi/docker-emscripten-crypt4gh all
+    elif [ "$1" = "emsdk" ]; then
+        /emsdk/upstream/emscripten/emmake make all
+    fi
+}
+
+if [ -x "/emsdk/emsdk" ]; then
+    echo "Running in an environment with natively supported emsdk. Building WASM contents using that."
+    export EMCC_CFLAGS="-I/emsdk/upstream/include -L/emsdk/upstream/lib -sINITIAL_MEMORY=26214400"
+    export EMCC_FORCE_STDLIBS="libc"
+
+    cd $WASM_ROOT
+
+    /emsdk/upstream/emscripten/emmake make clean
+    buildWasm "emsdk"
+
     cd $SCRIPT_ROOT
-    npx webpack --config "${WASM_ROOT}/wasm-webpack.config.js"
-)
+else
+    echo "emsdk was not configured, checking for docker"
+    [ ! -x "$(command -v docker)" ] && echo "Docker is used to build the javascript WebAssembly dependencies, but it's not installed." && exit;
+    if ! docker version > /dev/null 2>&1; then echo "Docker is installed, but it seems like there's an error."; exit 1; fi
 
-docker run --rm -it --mount type=bind,source="${WASM_ROOT}",target=/src/ ghcr.io/cscfi/docker-emscripten-crypt4gh all
+    docker run --rm -it --mount type=bind,source="${WASM_ROOT}",target=/src/ ghcr.io/cscfi/docker-emscripten-crypt4gh clean
+    buildWasm "docker"
+fi
 
 cp "${WASM_ROOT}/build/upworker.js" "${SCRIPT_ROOT}/public/"
 cp "${WASM_ROOT}/build/downworker.js" "${SCRIPT_ROOT}/public/"
@@ -25,3 +47,16 @@ cp "${WASM_ROOT}/build/upworker-post.js.map" "${SCRIPT_ROOT}/public/"
 cp "${WASM_ROOT}/build/downworker-post.js.map" "${SCRIPT_ROOT}/public/"
 cp "${WASM_ROOT}/build/upworker.wasm" "${SCRIPT_ROOT}/public/"
 cp "${WASM_ROOT}/build/downworker.wasm" "${SCRIPT_ROOT}/public/"
+
+# Copy over the S3 compatible versions as well
+cp "${WASM_ROOT}/build/s3upworker.js" "${SCRIPT_ROOT}/public/"
+cp "${WASM_ROOT}/build/s3upworker.wasm" "${SCRIPT_ROOT}/public/"
+cp "${WASM_ROOT}/build/crypt-post-s3upload.js.map" "${SCRIPT_ROOT}/public/"
+
+cp "${WASM_ROOT}/build/s3downworker.js" "${SCRIPT_ROOT}/public/"
+cp "${WASM_ROOT}/build/s3downworker.wasm" "${SCRIPT_ROOT}/public/"
+cp "${WASM_ROOT}/build/crypt-post-s3download.js.map" "${SCRIPT_ROOT}/public/"
+
+cp "${WASM_ROOT}/build/s3headerworker.js" "${SCRIPT_ROOT}/public/"
+cp "${WASM_ROOT}/build/s3headerworker.wasm" "${SCRIPT_ROOT}/public/"
+cp "${WASM_ROOT}/build/crypt-post-headers.js.map" "${SCRIPT_ROOT}/public/"
